requirements.txt
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.0
openai>=1.30.0
google-genai>=1.0.0
tenacity>=8.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.35.0
st-star-rating>=0.0.6
pandas>=2.0.0
rich>=13.7.0
requests>=2.31.0
